rename saveGuestListCurrent to saveGuestList to be consistent with function name; move to src directory
